test: test complex cases on all drivers too
